+++ libs/langchain-openai/.eslintrc.cjs
@@ -65,7 +65,6 @@ module.exports = {
     "no-lonely-if": 0,
     "prefer-rest-params": 0,
     "new-cap": ["error", { properties: false, capIsNew: false }],
-    'jest/no-focused-tests': 'error',
   },
   overrides: [
     {
