--- libs/langchain-openai/jest.config.cjs
@@ -1,21 +1,0 @@
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-  preset: "ts-jest/presets/default-esm",
-  testEnvironment: "./jest.env.cjs",
-  modulePathIgnorePatterns: ["dist/", "docs/"],
-  moduleNameMapper: {
-    "^(\\.{1,2}/.*)\\.js$": "$1",
-  },
-  transform: {
-    '^.+\\.tsx?$': ['@swc/jest'],
-  },
-  transformIgnorePatterns: [
-    "/node_modules/",
-    "\\.pnp\\.[^\\/]+$",
-    "./scripts/jest-setup-after-env.js",
-  ],
-  setupFiles: ["dotenv/config"],
-  setupFilesAfterEnv: ["./scripts/jest-setup-after-env.js"],
-  testTimeout: 20_000,
-  passWithNoTests: true
-};
