+++ libs/langchain-openai/package.json
@@ -23,12 +23,12 @@
     "lint:fix": "yarn lint:eslint --fix && yarn lint:dpdm",
     "clean": "rm -rf .turbo dist/",
     "prepack": "yarn build",
-    "test": "NODE_OPTIONS=--experimental-vm-modules jest --testPathIgnorePatterns=\\.int\\.test.ts --testTimeout 30000 --maxWorkers=50%",
-    "test:watch": "NODE_OPTIONS=--experimental-vm-modules jest --watch --testPathIgnorePatterns=\\.int\\.test.ts",
-    "test:single": "NODE_OPTIONS=--experimental-vm-modules yarn run jest --config jest.config.cjs --testTimeout 100000",
-    "test:int": "NODE_OPTIONS=--experimental-vm-modules jest --testPathPattern=\\.int\\.test.ts --testTimeout 100000 --maxWorkers=50%",
-    "test:standard:unit": "NODE_OPTIONS=--experimental-vm-modules jest --testPathPattern=\\.standard\\.test.ts --testTimeout 100000 --maxWorkers=50%",
-    "test:standard:int": "NODE_OPTIONS=--experimental-vm-modules jest --testPathPattern=\\.standard\\.int\\.test.ts --testTimeout 100000 --maxWorkers=50%",
+    "test": "vitest run --exclude=\"src/**/*.int.test.ts\" --testTimeout 30000",
+    "test:watch": "vitest --exclude=\"src/**/*.int.test.ts\"",
+    "test:single": "vitest run --testTimeout 100000",
+    "test:int": "vitest run --include=\"src/**/*.int.test.ts\" --testTimeout 100000",
+    "test:standard:unit": "vitest run --include=\"src/**/*.standard.test.ts\" --testTimeout 100000",
+    "test:standard:int": "vitest run --include=\"src/**/*.standard.int.test.ts\" --testTimeout 100000",
     "test:standard": "yarn test:standard:unit && yarn test:standard:int",
     "format": "prettier --config .prettierrc --write \"src\"",
     "format:check": "prettier --config .prettierrc --check \"src\""
@@ -46,27 +46,22 @@
   },
   "devDependencies": {
     "@azure/identity": "^4.2.1",
-    "@jest/globals": "^29.5.0",
     "@langchain/core": "workspace:*",
     "@langchain/scripts": ">=0.1.0 <0.2.0",
     "@langchain/standard-tests": "0.0.0",
     "@swc/core": "^1.3.90",
-    "@swc/jest": "^0.2.29",
     "dpdm": "^3.12.0",
     "eslint": "^8.33.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jest": "^27.6.0",
     "eslint-plugin-no-instanceof": "^1.0.1",
     "eslint-plugin-prettier": "^4.2.1",
-    "jest": "^29.5.0",
-    "jest-environment-node": "^29.6.4",
     "prettier": "^2.8.3",
     "release-it": "^18.1.2",
     "rimraf": "^5.0.1",
-    "ts-jest": "^29.1.0",
-    "typescript": "~5.1.6"
+    "typescript": "~5.1.6",
+    "vitest": "^3.2.3"
   },
   "publishConfig": {
     "access": "public"
@@ -87,19 +82,21 @@
   ],
   "exports": {
     ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
+      "types": {
+        "import": "./index.d.ts",
+        "require": "./index.d.cts",
+        "default": "./index.d.ts"
       },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
-      },
-      "input": "./src/index.ts"
+      "import": "./index.js",
+      "require": "./index.cjs"
     },
     "./package.json": "./package.json"
   },
   "files": [
-    "dist/"
+    "dist/",
+    "index.cjs",
+    "index.js",
+    "index.d.ts",
+    "index.d.cts"
   ]
 }
