+++ libs/langchain-openai/src/tests/azure/chat_models.standard.int.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-process-env */
-import { test, expect, afterAll } from "@jest/globals";
+import { test, expect, afterAll, vi } from "vitest";
 import { ChatModelIntegrationTests } from "@langchain/standard-tests";
 import { AIMessageChunk } from "@langchain/core/messages";
 import { AzureChatOpenAI } from "../../azure/chat_models.js";
