+++ libs/langchain-openai/src/tests/azure/chat_models.standard.test.ts
@@ -1,9 +1,16 @@
 /* eslint-disable no-process-env */
-import { test, expect } from "@jest/globals";
+import { test, expect, vi } from "vitest";
 import { ChatModelUnitTests } from "@langchain/standard-tests";
 import { AIMessageChunk } from "@langchain/core/messages";
 import { AzureChatOpenAI } from "../../azure/chat_models.js";
 import { ChatOpenAICallOptions } from "../../chat_models.js";
+
+/**
+ * Mock the @jest/globals module to use the vitest test and expect functions.
+ * This is necessary because the @langchain/standard-tests package uses the
+ * @jest/globals module to run the tests.
+ */
+vi.mock('@jest/globals', () => ({ test, expect }));
 
 class AzureChatOpenAIStandardUnitTests extends ChatModelUnitTests<
   ChatOpenAICallOptions,
