+++ libs/langchain-openai/src/tests/azure/chat_models.test.ts
@@ -1,3 +1,4 @@
+import { test, expect, beforeEach } from "vitest";
 /* eslint-disable no-process-env */
 import { AzureChatOpenAI } from "../../azure/chat_models.js";
