+++ libs/langchain-openai/src/tests/chat_models.standard.int.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-process-env */
-import { test } from "@jest/globals";
+import { test, vi } from "vitest";
 import { ChatModelIntegrationTests } from "@langchain/standard-tests";
 import { AIMessage, AIMessageChunk } from "@langchain/core/messages";
 import { readFileSync } from "fs";
