+++ libs/langchain-openai/src/tests/chat_models_responses.int.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-process-env */
-import { test, expect } from "@jest/globals";
+import { test, expect } from "vitest";
 import {
   AIMessage,
   AIMessageChunk,
