+++ libs/langchain-openai/src/tests/chat_models_structured_output.int.test.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 import { zodToJsonSchema } from "zod-to-json-schema";
 import { ChatPromptTemplate } from "@langchain/core/prompts";
 import { AIMessage, AIMessageChunk } from "@langchain/core/messages";
-import { test, expect, describe, it } from "@jest/globals";
+import { test, expect, describe, it } from "vitest";
 import { concat } from "@langchain/core/utils/stream";
 import { ChatOpenAI } from "../chat_models.js";
