+++ libs/langchain-openai/src/tests/embeddings.int.test.ts
@@ -1,4 +1,4 @@
-import { test, expect } from "@jest/globals";
+import { test, expect } from "vitest";
 import { OpenAIEmbeddings } from "../embeddings.js";
 
 test("Test OpenAIEmbeddings.embedQuery", async () => {
